Move gen_in_network_links() to helpers.py
